+++ test/index.test.ts
@@ -64,6 +64,9 @@ describe('Swagger', () => {
 
         const res = await app.handle(req('/v2/swagger'))
         expect(res.status).toBe(200)
+
+        const resJson = await app.handle(req('/v2/swagger/json'))
+        expect(resJson.status).toBe(200)
     })
 
     it('Swagger UI options', async () => {
@@ -75,56 +78,70 @@ describe('Swagger', () => {
             })
         )
         const res = await app.handle(req('/swagger')).then((x) => x.text())
-        const expected = `
-        window.onload = () => {
-            window.ui = SwaggerUIBundle({"url":"/swagger/json","dom_id":"#swagger-ui","persistAuthorization":true});
-        };
-        `
+        const expected = `"persistAuthorization":true`
+
         expect(res.trim().includes(expected.trim())).toBe(true)
     })
 
     it('should not return content response when using Void type', async () => {
-        const app = new Elysia().use(
-            swagger())
-            .get('/void', () => {}, {
-                response: { 204: t.Void({
+        const app = new Elysia().use(swagger()).get('/void', () => {}, {
+            response: {
+                204: t.Void({
                     description: 'Void response'
-            })}});
+                })
+            }
+        })
 
         const res = await app.handle(req('/swagger/json'))
         expect(res.status).toBe(200)
-        const response = await res.json();
-        expect(response.paths['/void'].get.responses['204'].description).toBe('Void response');
-        expect(response.paths['/void'].get.responses['204'].content).toBeUndefined();
+        const response = await res.json()
+        expect(response.paths['/void'].get.responses['204'].description).toBe(
+            'Void response'
+        )
+        expect(
+            response.paths['/void'].get.responses['204'].content
+        ).toBeUndefined()
     })
 
     it('should not return content response when using Undefined type', async () => {
-        const app = new Elysia().use(
-            swagger())
+        const app = new Elysia()
+            .use(swagger())
             .get('/undefined', () => undefined, {
-                response: { 204: t.Undefined({
-                    description: 'Undefined response'
-            })}});
+                response: {
+                    204: t.Undefined({
+                        description: 'Undefined response'
+                    })
+                }
+            })
 
         const res = await app.handle(req('/swagger/json'))
         expect(res.status).toBe(200)
-        const response = await res.json();
-        expect(response.paths['/undefined'].get.responses['204'].description).toBe('Undefined response');
-        expect(response.paths['/undefined'].get.responses['204'].content).toBeUndefined();
+        const response = await res.json()
+        expect(
+            response.paths['/undefined'].get.responses['204'].description
+        ).toBe('Undefined response')
+        expect(
+            response.paths['/undefined'].get.responses['204'].content
+        ).toBeUndefined()
     })
 
     it('should not return content response when using Null type', async () => {
-        const app = new Elysia().use(
-            swagger())
-            .get('/null', () => null, {
-                response: { 204: t.Null({
+        const app = new Elysia().use(swagger()).get('/null', () => null, {
+            response: {
+                204: t.Null({
                     description: 'Null response'
-            })}});
+                })
+            }
+        })
 
         const res = await app.handle(req('/swagger/json'))
         expect(res.status).toBe(200)
-        const response = await res.json();
-        expect(response.paths['/null'].get.responses['204'].description).toBe('Null response');
-        expect(response.paths['/null'].get.responses['204'].content).toBeUndefined();
+        const response = await res.json()
+        expect(response.paths['/null'].get.responses['204'].description).toBe(
+            'Null response'
+        )
+        expect(
+            response.paths['/null'].get.responses['204'].content
+        ).toBeUndefined()
     })
 })